pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stratifyai"
version = "0.1.0"
description = "Unified multi-provider LLM abstraction module with intelligent routing, cost tracking, and caching"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Steven Cotton", email = "cototnbytes@gmail.com"}
]
maintainers = [
    {name = "Steven Cotton", email = "cototnbytes@gmail.com"}
]
license = {text = "MIT"}
keywords = [
    "llm",
    "openai",
    "anthropic",
    "google",
    "gemini",
    "ai",
    "chatgpt",
    "claude",
    "deepseek",
    "groq",
    "grok",
    "ollama",
    "multi-provider",
    "abstraction",
    "routing",
    "cost-tracking",
    "caching"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "google-genai>=1.0.0",
    "aioboto3>=12.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
]
rag = [
    "chromadb>=0.5.0",
]
all = [
    "stratifyai[dev,cli,web,rag]",
]

[project.urls]
Homepage = "https://github.com/Bytes0211/stratifyai"
Repository = "https://github.com/Bytes0211/stratifyai"
Documentation = "https://github.com/Bytes0211/stratifyai#readme"
"Bug Tracker" = "https://github.com/Bytes0211/stratifyai/issues"

[project.scripts]
stratifyai = "cli.stratifyai_cli:main"

[tool.setuptools]
packages = ["stratifyai", "stratifyai.chat", "stratifyai.providers", "stratifyai.utils", "cli"]

[tool.setuptools.package-data]
stratifyai = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "asyncio: marks tests as async tests",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
